requirements.txt
selenium==4.11.2
gunicorn==20.0.4
flask==2.1.0
imap-tools
azure.storage.queue
azure.storage.blob
cryptography==3.4.6
easyimap	
lxml
beautifulsoup4
pandas==1.5.3
pdfservices-sdk==2.3.0
requests
PyPDF2
msal
openpyxl
pytz
azure.data.tables
tabula-py
usaddress
Pillow
img2pdf
PyMuPDF==1.22.5